+++ datapipe/step/batch_transform.py
@@ -105,7 +105,7 @@ class TransformMetaTable:
         self.primary_schema = primary_schema
         self.primary_keys = [i.name for i in primary_schema]
 
-        self.sql_schema = [i.copy() for i in primary_schema + TRANSFORM_META_SCHEMA]  # type: ignore
+        self.sql_schema = [i._copy() for i in primary_schema + TRANSFORM_META_SCHEMA]  # type: ignore
 
         self.sql_table = Table(
             name,
@@ -116,6 +116,13 @@ class TransformMetaTable:
 
         if create_table:
             self.sql_table.create(self.dbconn.con, checkfirst=True)
+
+    def __reduce__(self) -> Tuple[Any, ...]:
+        return self.__class__, (
+            self.dbconn,
+            self.name,
+            self.primary_schema,
+        )
 
     def insert_rows(
         self,
@@ -228,8 +235,9 @@ class TransformMetaTable:
         Получить количество строк метаданных трансформации.
         """
 
-        sql = select([func.count()]).select_from(self.sql_table)
-        res = self.dbconn.con.execute(sql).fetchone()
+        sql = select(func.count()).select_from(self.sql_table)
+        with self.dbconn.con.begin() as con:
+            res = con.execute(sql).fetchone()
 
         assert res is not None and len(res) == 1
         return res[0]
@@ -385,17 +393,17 @@ class BaseBatchTransformStep(ComputeStep):
                         func.coalesce(*[cte.c[key] for cte in ctes_with_key]).label(key)
                     )
 
-            agg = ds.meta_dbconn.func_greatest(
-                *[subq.c[agg_col] for (_, subq) in ctes]
+            agg = func.max(
+                ds.meta_dbconn.func_greatest(*[subq.c[agg_col] for (_, subq) in ctes])
             ).label(agg_col)
 
             _, first_cte = ctes[0]
 
-            sql = select(*coalesce_keys + [agg]).distinct().select_from(first_cte)
+            sql = select(*coalesce_keys + [agg]).select_from(first_cte)
 
             for _, cte in ctes[1:]:
                 if len(common_transform_keys) > 0:
-                    sql = sql.join(
+                    sql = sql.outerjoin(
                         cte,
                         onclause=and_(
                             *[
@@ -406,10 +414,13 @@ class BaseBatchTransformStep(ComputeStep):
                         full=True,
                     )
                 else:
-                    sql = sql.join(
+                    sql = sql.outerjoin(
                         cte,
                         onclause=literal(True),
+                        full=True,
                     )
+
+            sql = sql.group_by(*coalesce_keys)
 
             return sql.cte(name=f"all__{agg_col}")
 
@@ -421,6 +432,11 @@ class BaseBatchTransformStep(ComputeStep):
             )
             for tbl in self.input_dts
         ]
+
+        # Filter out ctes that do not have intersection with transform keys.
+        # These ctes do not affect the result, but may dramatically increase
+        # size of the temporary table during query execution.
+        inp_ctes = [(keys, cte) for (keys, cte) in inp_ctes if len(keys) > 0]
 
         inp = _make_agg_of_agg(inp_ctes, "update_ts")
 
@@ -472,12 +488,12 @@ class BaseBatchTransformStep(ComputeStep):
         else:
             if order == "desc":
                 sql = sql.order_by(
-                    desc(*[column(k) for k in order_by]),
+                    *[desc(column(k)) for k in order_by],
                     out.c.priority.desc().nullslast(),
                 )
             elif order == "asc":
                 sql = sql.order_by(
-                    asc(*[column(k) for k in order_by]),
+                    *[asc(column(k)) for k in order_by],
                     out.c.priority.desc().nullslast(),
                 )
         return (self.transform_keys, sql)
@@ -512,7 +528,7 @@ class BaseBatchTransformStep(ComputeStep):
 
         with ds.meta_dbconn.con.begin() as con:
             idx_count = con.execute(
-                select([func.count()]).select_from(
+                select(*[func.count()]).select_from(
                     alias(sql.subquery(), name="union_select")
                 )
             ).scalar()